OoT cutscenes as first-class objects (#18)
* Changed cutscenes to being a first class (empty type) object, separate cutscene exporting

* Moved some classes to new files to fix circular import issues

* Added ability to save other cutscenes not in headers into scene file

* Made most changes according to Dragorn421's suggestions

* Fixed a typo
