+++ fast64_internal/oot/oot_constants.py
@@ -58,434 +58,434 @@ ootEnumTransitionActorID = [
 ootEnumActorID = [
 	("Custom", "Custom", "Custom"),
 	("ACTOR_PLAYER", "PLAYER", "PLAYER"),
-    ("ACTOR_EN_TEST", "EN_TEST", "EN_TEST"),
-    ("ACTOR_EN_GIRLA", "EN_GIRLA", "EN_GIRLA"),
-    ("ACTOR_EN_PART", "EN_PART", "EN_PART"),
-    ("ACTOR_EN_LIGHT", "EN_LIGHT", "EN_LIGHT"),
-    ("ACTOR_EN_DOOR", "EN_DOOR", "EN_DOOR"),
-    ("ACTOR_EN_BOX", "EN_BOX", "EN_BOX"),
-    ("ACTOR_BG_DY_YOSEIZO", "BG_DY_YOSEIZO", "BG_DY_YOSEIZO"),
-    ("ACTOR_BG_HIDAN_FIREWALL", "BG_HIDAN_FIREWALL", "BG_HIDAN_FIREWALL"),
-    ("ACTOR_EN_POH", "EN_POH", "EN_POH"),
-    ("ACTOR_EN_OKUTA", "EN_OKUTA", "EN_OKUTA"),
-    ("ACTOR_BG_YDAN_SP", "BG_YDAN_SP", "BG_YDAN_SP"),
-    ("ACTOR_EN_BOM", "EN_BOM", "EN_BOM"),
-    ("ACTOR_EN_WALLMAS", "EN_WALLMAS", "EN_WALLMAS"),
-    ("ACTOR_EN_DODONGO", "EN_DODONGO", "EN_DODONGO"),
-    ("ACTOR_EN_FIREFLY", "EN_FIREFLY", "EN_FIREFLY"),
-    ("ACTOR_EN_HORSE", "EN_HORSE", "EN_HORSE"),
-    ("ACTOR_EN_ITEM00", "EN_ITEM00", "EN_ITEM00"),
-    ("ACTOR_EN_ARROW", "EN_ARROW", "EN_ARROW"),
-    ("ACTOR_EN_ELF", "EN_ELF", "EN_ELF"),
-    ("ACTOR_EN_NIW", "EN_NIW", "EN_NIW"),
-    ("ACTOR_EN_TITE", "EN_TITE", "EN_TITE"),
-    ("ACTOR_EN_REEBA", "EN_REEBA", "EN_REEBA"),
-    ("ACTOR_EN_PEEHAT", "EN_PEEHAT", "EN_PEEHAT"),
-    ("ACTOR_EN_BUTTE", "EN_BUTTE", "EN_BUTTE"),
-    ("ACTOR_EN_INSECT", "EN_INSECT", "EN_INSECT"),
-    ("ACTOR_EN_FISH", "EN_FISH", "EN_FISH"),
-    ("ACTOR_EN_HOLL", "EN_HOLL", "EN_HOLL"),
-    ("ACTOR_EN_SCENE_CHANGE", "EN_SCENE_CHANGE", "EN_SCENE_CHANGE"),
-    ("ACTOR_EN_ZF", "EN_ZF", "EN_ZF"),
-    ("ACTOR_EN_HATA", "EN_HATA", "EN_HATA"),
-    ("ACTOR_BOSS_DODONGO", "BOSS_DODONGO", "BOSS_DODONGO"),
-    ("ACTOR_BOSS_GOMA", "BOSS_GOMA", "BOSS_GOMA"),
-    ("ACTOR_EN_ZL1", "EN_ZL1", "EN_ZL1"),
-    ("ACTOR_EN_VIEWER", "EN_VIEWER", "EN_VIEWER"),
-    ("ACTOR_EN_GOMA", "EN_GOMA", "EN_GOMA"),
-    ("ACTOR_BG_PUSHBOX", "BG_PUSHBOX", "BG_PUSHBOX"),
-    ("ACTOR_EN_BUBBLE", "EN_BUBBLE", "EN_BUBBLE"),
-    ("ACTOR_DOOR_SHUTTER", "DOOR_SHUTTER", "DOOR_SHUTTER"),
-    ("ACTOR_EN_DODOJR", "EN_DODOJR", "EN_DODOJR"),
-    ("ACTOR_EN_BDFIRE", "EN_BDFIRE", "EN_BDFIRE"),
-    ("ACTOR_EN_BOOM", "EN_BOOM", "EN_BOOM"),
-    ("ACTOR_EN_TORCH2", "EN_TORCH2", "EN_TORCH2"),
-    ("ACTOR_EN_BILI", "EN_BILI", "EN_BILI"),
-    ("ACTOR_EN_TP", "EN_TP", "EN_TP"),
-    ("ACTOR_EN_ST", "EN_ST", "EN_ST"),
-    ("ACTOR_EN_BW", "EN_BW", "EN_BW"),
-    ("ACTOR_EN_A_OBJ", "EN_A_OBJ", "EN_A_OBJ"),
-    ("ACTOR_EN_EIYER", "EN_EIYER", "EN_EIYER"),
-    ("ACTOR_EN_RIVER_SOUND", "EN_RIVER_SOUND", "EN_RIVER_SOUND"),
-    ("ACTOR_EN_HORSE_NORMAL", "EN_HORSE_NORMAL", "EN_HORSE_NORMAL"),
-    ("ACTOR_EN_OSSAN", "EN_OSSAN", "EN_OSSAN"),
-    ("ACTOR_BG_TREEMOUTH", "BG_TREEMOUTH", "BG_TREEMOUTH"),
-    ("ACTOR_BG_DODOAGO", "BG_DODOAGO", "BG_DODOAGO"),
-    ("ACTOR_BG_HIDAN_DALM", "BG_HIDAN_DALM", "BG_HIDAN_DALM"),
-    ("ACTOR_BG_HIDAN_HROCK", "BG_HIDAN_HROCK", "BG_HIDAN_HROCK"),
-    ("ACTOR_EN_HORSE_GANON", "EN_HORSE_GANON", "EN_HORSE_GANON"),
-    ("ACTOR_BG_HIDAN_ROCK", "BG_HIDAN_ROCK", "BG_HIDAN_ROCK"),
-    ("ACTOR_BG_HIDAN_RSEKIZOU", "BG_HIDAN_RSEKIZOU", "BG_HIDAN_RSEKIZOU"),
-    ("ACTOR_BG_HIDAN_SEKIZOU", "BG_HIDAN_SEKIZOU", "BG_HIDAN_SEKIZOU"),
-    ("ACTOR_BG_HIDAN_SIMA", "BG_HIDAN_SIMA", "BG_HIDAN_SIMA"),
-    ("ACTOR_BG_HIDAN_SYOKU", "BG_HIDAN_SYOKU", "BG_HIDAN_SYOKU"),
-    ("ACTOR_EN_XC", "EN_XC", "EN_XC"),
-    ("ACTOR_BG_HIDAN_CURTAIN", "BG_HIDAN_CURTAIN", "BG_HIDAN_CURTAIN"),
-    ("ACTOR_BG_SPOT00_HANEBASI", "BG_SPOT00_HANEBASI", "BG_SPOT00_HANEBASI"),
-    ("ACTOR_EN_MB", "EN_MB", "EN_MB"),
-    ("ACTOR_EN_BOMBF", "EN_BOMBF", "EN_BOMBF"),
-    ("ACTOR_EN_ZL2", "EN_ZL2", "EN_ZL2"),
-    ("ACTOR_BG_HIDAN_FSLIFT", "BG_HIDAN_FSLIFT", "BG_HIDAN_FSLIFT"),
-    ("ACTOR_EN_OE2", "EN_OE2", "EN_OE2"),
-    ("ACTOR_BG_YDAN_HASI", "BG_YDAN_HASI", "BG_YDAN_HASI"),
-    ("ACTOR_BG_YDAN_MARUTA", "BG_YDAN_MARUTA", "BG_YDAN_MARUTA"),
-    ("ACTOR_BOSS_GANONDROF", "BOSS_GANONDROF", "BOSS_GANONDROF"),
-    ("ACTOR_EN_AM", "EN_AM", "EN_AM"),
-    ("ACTOR_EN_DEKUBABA", "EN_DEKUBABA", "EN_DEKUBABA"),
-    ("ACTOR_EN_M_FIRE1", "EN_M_FIRE1", "EN_M_FIRE1"),
-    ("ACTOR_EN_M_THUNDER", "EN_M_THUNDER", "EN_M_THUNDER"),
-    ("ACTOR_BG_DDAN_JD", "BG_DDAN_JD", "BG_DDAN_JD"),
-    ("ACTOR_BG_BREAKWALL", "BG_BREAKWALL", "BG_BREAKWALL"),
-    ("ACTOR_EN_JJ", "EN_JJ", "EN_JJ"),
-    ("ACTOR_EN_HORSE_ZELDA", "EN_HORSE_ZELDA", "EN_HORSE_ZELDA"),
-    ("ACTOR_BG_DDAN_KD", "BG_DDAN_KD", "BG_DDAN_KD"),
-    ("ACTOR_DOOR_WARP1", "DOOR_WARP1", "DOOR_WARP1"),
-    ("ACTOR_OBJ_SYOKUDAI", "OBJ_SYOKUDAI", "OBJ_SYOKUDAI"),
-    ("ACTOR_ITEM_B_HEART", "ITEM_B_HEART", "ITEM_B_HEART"),
-    ("ACTOR_EN_DEKUNUTS", "EN_DEKUNUTS", "EN_DEKUNUTS"),
-    ("ACTOR_BG_MENKURI_KAITEN", "BG_MENKURI_KAITEN", "BG_MENKURI_KAITEN"),
-    ("ACTOR_BG_MENKURI_EYE", "BG_MENKURI_EYE", "BG_MENKURI_EYE"),
-    ("ACTOR_EN_VALI", "EN_VALI", "EN_VALI"),
-    ("ACTOR_BG_MIZU_MOVEBG", "BG_MIZU_MOVEBG", "BG_MIZU_MOVEBG"),
-    ("ACTOR_BG_MIZU_WATER", "BG_MIZU_WATER", "BG_MIZU_WATER"),
-    ("ACTOR_ARMS_HOOK", "ARMS_HOOK", "ARMS_HOOK"),
-    ("ACTOR_EN_FHG", "EN_FHG", "EN_FHG"),
-    ("ACTOR_BG_MORI_HINERI", "BG_MORI_HINERI", "BG_MORI_HINERI"),
-    ("ACTOR_EN_BB", "EN_BB", "EN_BB"),
-    ("ACTOR_BG_TOKI_HIKARI", "BG_TOKI_HIKARI", "BG_TOKI_HIKARI"),
-    ("ACTOR_EN_YUKABYUN", "EN_YUKABYUN", "EN_YUKABYUN"),
-    ("ACTOR_BG_TOKI_SWD", "BG_TOKI_SWD", "BG_TOKI_SWD"),
-    ("ACTOR_EN_FHG_FIRE", "EN_FHG_FIRE", "EN_FHG_FIRE"),
-    ("ACTOR_BG_MJIN", "BG_MJIN", "BG_MJIN"),
-    ("ACTOR_BG_HIDAN_KOUSI", "BG_HIDAN_KOUSI", "BG_HIDAN_KOUSI"),
-    ("ACTOR_DOOR_TOKI", "DOOR_TOKI", "DOOR_TOKI"),
-    ("ACTOR_BG_HIDAN_HAMSTEP", "BG_HIDAN_HAMSTEP", "BG_HIDAN_HAMSTEP"),
-    ("ACTOR_EN_BIRD", "EN_BIRD", "EN_BIRD"),
-    ("ACTOR_EN_WOOD02", "EN_WOOD02", "EN_WOOD02"),
-    ("ACTOR_EN_LIGHTBOX", "EN_LIGHTBOX", "EN_LIGHTBOX"),
-    ("ACTOR_EN_PU_BOX", "EN_PU_BOX", "EN_PU_BOX"),
-    ("ACTOR_EN_TRAP", "EN_TRAP", "EN_TRAP"),
-    ("ACTOR_EN_AROW_TRAP", "EN_AROW_TRAP", "EN_AROW_TRAP"),
-    ("ACTOR_EN_VASE", "EN_VASE", "EN_VASE"),
-    ("ACTOR_EN_TA", "EN_TA", "EN_TA"),
-    ("ACTOR_EN_TK", "EN_TK", "EN_TK"),
-    ("ACTOR_BG_MORI_BIGST", "BG_MORI_BIGST", "BG_MORI_BIGST"),
-    ("ACTOR_BG_MORI_ELEVATOR", "BG_MORI_ELEVATOR", "BG_MORI_ELEVATOR"),
-    ("ACTOR_BG_MORI_KAITENKABE", "BG_MORI_KAITENKABE", "BG_MORI_KAITENKABE"),
-    ("ACTOR_BG_MORI_RAKKATENJO", "BG_MORI_RAKKATENJO", "BG_MORI_RAKKATENJO"),
-    ("ACTOR_EN_VM", "EN_VM", "EN_VM"),
-    ("ACTOR_DEMO_EFFECT", "DEMO_EFFECT", "DEMO_EFFECT"),
-    ("ACTOR_DEMO_KANKYO", "DEMO_KANKYO", "DEMO_KANKYO"),
-    ("ACTOR_BG_HIDAN_FWBIG", "BG_HIDAN_FWBIG", "BG_HIDAN_FWBIG"),
-    ("ACTOR_EN_FLOORMAS", "EN_FLOORMAS", "EN_FLOORMAS"),
-    ("ACTOR_EN_HEISHI1", "EN_HEISHI1", "EN_HEISHI1"),
-    ("ACTOR_EN_RD", "EN_RD", "EN_RD"),
-    ("ACTOR_EN_PO_SISTERS", "EN_PO_SISTERS", "EN_PO_SISTERS"),
-    ("ACTOR_BG_HEAVY_BLOCK", "BG_HEAVY_BLOCK", "BG_HEAVY_BLOCK"),
-    ("ACTOR_BG_PO_EVENT", "BG_PO_EVENT", "BG_PO_EVENT"),
-    ("ACTOR_OBJ_MURE", "OBJ_MURE", "OBJ_MURE"),
-    ("ACTOR_EN_SW", "EN_SW", "EN_SW"),
-    ("ACTOR_BOSS_FD", "BOSS_FD", "BOSS_FD"),
-    ("ACTOR_OBJECT_KANKYO", "OBJECT_KANKYO", "OBJECT_KANKYO"),
-    ("ACTOR_EN_DU", "EN_DU", "EN_DU"),
-    ("ACTOR_EN_FD", "EN_FD", "EN_FD"),
-    ("ACTOR_EN_HORSE_LINK_CHILD", "EN_HORSE_LINK_CHILD", "EN_HORSE_LINK_CHILD"),
-    ("ACTOR_DOOR_ANA", "DOOR_ANA", "DOOR_ANA"),
-    ("ACTOR_BG_SPOT02_OBJECTS", "BG_SPOT02_OBJECTS", "BG_SPOT02_OBJECTS"),
-    ("ACTOR_BG_HAKA", "BG_HAKA", "BG_HAKA"),
-    ("ACTOR_MAGIC_WIND", "MAGIC_WIND", "MAGIC_WIND"),
-    ("ACTOR_MAGIC_FIRE", "MAGIC_FIRE", "MAGIC_FIRE"),
-    ("ACTOR_EN_RU1", "EN_RU1", "EN_RU1"),
-    ("ACTOR_BOSS_FD2", "BOSS_FD2", "BOSS_FD2"),
-    ("ACTOR_EN_FD_FIRE", "EN_FD_FIRE", "EN_FD_FIRE"),
-    ("ACTOR_EN_DH", "EN_DH", "EN_DH"),
-    ("ACTOR_EN_DHA", "EN_DHA", "EN_DHA"),
-    ("ACTOR_EN_RL", "EN_RL", "EN_RL"),
-    ("ACTOR_EN_ENCOUNT1", "EN_ENCOUNT1", "EN_ENCOUNT1"),
-    ("ACTOR_DEMO_DU", "DEMO_DU", "DEMO_DU"),
-    ("ACTOR_DEMO_IM", "DEMO_IM", "DEMO_IM"),
-    ("ACTOR_DEMO_TRE_LGT", "DEMO_TRE_LGT", "DEMO_TRE_LGT"),
-    ("ACTOR_EN_FW", "EN_FW", "EN_FW"),
-    ("ACTOR_BG_VB_SIMA", "BG_VB_SIMA", "BG_VB_SIMA"),
-    ("ACTOR_EN_VB_BALL", "EN_VB_BALL", "EN_VB_BALL"),
-    ("ACTOR_BG_HAKA_MEGANE", "BG_HAKA_MEGANE", "BG_HAKA_MEGANE"),
-    ("ACTOR_BG_HAKA_MEGANEBG", "BG_HAKA_MEGANEBG", "BG_HAKA_MEGANEBG"),
-    ("ACTOR_BG_HAKA_SHIP", "BG_HAKA_SHIP", "BG_HAKA_SHIP"),
-    ("ACTOR_BG_HAKA_SGAMI", "BG_HAKA_SGAMI", "BG_HAKA_SGAMI"),
-    ("ACTOR_EN_HEISHI2", "EN_HEISHI2", "EN_HEISHI2"),
-    ("ACTOR_EN_ENCOUNT2", "EN_ENCOUNT2", "EN_ENCOUNT2"),
-    ("ACTOR_EN_FIRE_ROCK", "EN_FIRE_ROCK", "EN_FIRE_ROCK"),
-    ("ACTOR_EN_BROB", "EN_BROB", "EN_BROB"),
-    ("ACTOR_MIR_RAY", "MIR_RAY", "MIR_RAY"),
-    ("ACTOR_BG_SPOT09_OBJ", "BG_SPOT09_OBJ", "BG_SPOT09_OBJ"),
-    ("ACTOR_BG_SPOT18_OBJ", "BG_SPOT18_OBJ", "BG_SPOT18_OBJ"),
-    ("ACTOR_BOSS_VA", "BOSS_VA", "BOSS_VA"),
-    ("ACTOR_BG_HAKA_TUBO", "BG_HAKA_TUBO", "BG_HAKA_TUBO"),
-    ("ACTOR_BG_HAKA_TRAP", "BG_HAKA_TRAP", "BG_HAKA_TRAP"),
-    ("ACTOR_BG_HAKA_HUTA", "BG_HAKA_HUTA", "BG_HAKA_HUTA"),
-    ("ACTOR_BG_HAKA_ZOU", "BG_HAKA_ZOU", "BG_HAKA_ZOU"),
-    ("ACTOR_BG_SPOT17_FUNEN", "BG_SPOT17_FUNEN", "BG_SPOT17_FUNEN"),
-    ("ACTOR_EN_SYATEKI_ITM", "EN_SYATEKI_ITM", "EN_SYATEKI_ITM"),
-    ("ACTOR_EN_SYATEKI_MAN", "EN_SYATEKI_MAN", "EN_SYATEKI_MAN"),
-    ("ACTOR_EN_TANA", "EN_TANA", "EN_TANA"),
-    ("ACTOR_EN_NB", "EN_NB", "EN_NB"),
-    ("ACTOR_BOSS_MO", "BOSS_MO", "BOSS_MO"),
-    ("ACTOR_EN_SB", "EN_SB", "EN_SB"),
-    ("ACTOR_EN_BIGOKUTA", "EN_BIGOKUTA", "EN_BIGOKUTA"),
-    ("ACTOR_EN_KAREBABA", "EN_KAREBABA", "EN_KAREBABA"),
-    ("ACTOR_BG_BDAN_OBJECTS", "BG_BDAN_OBJECTS", "BG_BDAN_OBJECTS"),
-    ("ACTOR_DEMO_SA", "DEMO_SA", "DEMO_SA"),
-    ("ACTOR_DEMO_GO", "DEMO_GO", "DEMO_GO"),
-    ("ACTOR_EN_IN", "EN_IN", "EN_IN"),
-    ("ACTOR_EN_TR", "EN_TR", "EN_TR"),
-    ("ACTOR_BG_SPOT16_BOMBSTONE", "BG_SPOT16_BOMBSTONE", "BG_SPOT16_BOMBSTONE"),
-    ("ACTOR_BG_HIDAN_KOWARERUKABE", "BG_HIDAN_KOWARERUKABE", "BG_HIDAN_KOWARERUKABE"),
-    ("ACTOR_BG_BOMBWALL", "BG_BOMBWALL", "BG_BOMBWALL"),
-    ("ACTOR_BG_SPOT08_ICEBLOCK", "BG_SPOT08_ICEBLOCK", "BG_SPOT08_ICEBLOCK"),
-    ("ACTOR_EN_RU2", "EN_RU2", "EN_RU2"),
-    ("ACTOR_OBJ_DEKUJR", "OBJ_DEKUJR", "OBJ_DEKUJR"),
-    ("ACTOR_BG_MIZU_UZU", "BG_MIZU_UZU", "BG_MIZU_UZU"),
-    ("ACTOR_BG_SPOT06_OBJECTS", "BG_SPOT06_OBJECTS", "BG_SPOT06_OBJECTS"),
-    ("ACTOR_BG_ICE_OBJECTS", "BG_ICE_OBJECTS", "BG_ICE_OBJECTS"),
-    ("ACTOR_BG_HAKA_WATER", "BG_HAKA_WATER", "BG_HAKA_WATER"),
-    ("ACTOR_EN_MA2", "EN_MA2", "EN_MA2"),
-    ("ACTOR_EN_BOM_CHU", "EN_BOM_CHU", "EN_BOM_CHU"),
-    ("ACTOR_EN_HORSE_GAME_CHECK", "EN_HORSE_GAME_CHECK", "EN_HORSE_GAME_CHECK"),
-    ("ACTOR_BOSS_TW", "BOSS_TW", "BOSS_TW"),
-    ("ACTOR_EN_RR", "EN_RR", "EN_RR"),
-    ("ACTOR_EN_BA", "EN_BA", "EN_BA"),
-    ("ACTOR_EN_BX", "EN_BX", "EN_BX"),
-    ("ACTOR_EN_ANUBICE", "EN_ANUBICE", "EN_ANUBICE"),
-    ("ACTOR_EN_ANUBICE_FIRE", "EN_ANUBICE_FIRE", "EN_ANUBICE_FIRE"),
-    ("ACTOR_BG_MORI_HASHIGO", "BG_MORI_HASHIGO", "BG_MORI_HASHIGO"),
-    ("ACTOR_BG_MORI_HASHIRA4", "BG_MORI_HASHIRA4", "BG_MORI_HASHIRA4"),
-    ("ACTOR_BG_MORI_IDOMIZU", "BG_MORI_IDOMIZU", "BG_MORI_IDOMIZU"),
-    ("ACTOR_BG_SPOT16_DOUGHNUT", "BG_SPOT16_DOUGHNUT", "BG_SPOT16_DOUGHNUT"),
-    ("ACTOR_BG_BDAN_SWITCH", "BG_BDAN_SWITCH", "BG_BDAN_SWITCH"),
-    ("ACTOR_EN_MA1", "EN_MA1", "EN_MA1"),
-    ("ACTOR_BOSS_GANON", "BOSS_GANON", "BOSS_GANON"),
-    ("ACTOR_BOSS_SST", "BOSS_SST", "BOSS_SST"),
-    ("ACTOR_EN_NY", "EN_NY", "EN_NY"),
-    ("ACTOR_EN_FR", "EN_FR", "EN_FR"),
-    ("ACTOR_ITEM_SHIELD", "ITEM_SHIELD", "ITEM_SHIELD"),
-    ("ACTOR_BG_ICE_SHELTER", "BG_ICE_SHELTER", "BG_ICE_SHELTER"),
-    ("ACTOR_EN_ICE_HONO", "EN_ICE_HONO", "EN_ICE_HONO"),
-    ("ACTOR_ITEM_OCARINA", "ITEM_OCARINA", "ITEM_OCARINA"),
-    ("ACTOR_MAGIC_DARK", "MAGIC_DARK", "MAGIC_DARK"),
-    ("ACTOR_DEMO_6K", "DEMO_6K", "DEMO_6K"),
-    ("ACTOR_EN_ANUBICE_TAG", "EN_ANUBICE_TAG", "EN_ANUBICE_TAG"),
-    ("ACTOR_BG_HAKA_GATE", "BG_HAKA_GATE", "BG_HAKA_GATE"),
-    ("ACTOR_BG_SPOT15_SAKU", "BG_SPOT15_SAKU", "BG_SPOT15_SAKU"),
-    ("ACTOR_BG_JYA_GOROIWA", "BG_JYA_GOROIWA", "BG_JYA_GOROIWA"),
-    ("ACTOR_BG_JYA_ZURERUKABE", "BG_JYA_ZURERUKABE", "BG_JYA_ZURERUKABE"),
-    ("ACTOR_BG_JYA_COBRA", "BG_JYA_COBRA", "BG_JYA_COBRA"),
-    ("ACTOR_BG_JYA_KANAAMI", "BG_JYA_KANAAMI", "BG_JYA_KANAAMI"),
-    ("ACTOR_FISHING", "FISHING", "FISHING"),
-    ("ACTOR_OBJ_OSHIHIKI", "OBJ_OSHIHIKI", "OBJ_OSHIHIKI"),
-    ("ACTOR_BG_GATE_SHUTTER", "BG_GATE_SHUTTER", "BG_GATE_SHUTTER"),
-    ("ACTOR_EFF_DUST", "EFF_DUST", "EFF_DUST"),
-    ("ACTOR_BG_SPOT01_FUSYA", "BG_SPOT01_FUSYA", "BG_SPOT01_FUSYA"),
-    ("ACTOR_BG_SPOT01_IDOHASHIRA", "BG_SPOT01_IDOHASHIRA", "BG_SPOT01_IDOHASHIRA"),
-    ("ACTOR_BG_SPOT01_IDOMIZU", "BG_SPOT01_IDOMIZU", "BG_SPOT01_IDOMIZU"),
-    ("ACTOR_BG_PO_SYOKUDAI", "BG_PO_SYOKUDAI", "BG_PO_SYOKUDAI"),
-    ("ACTOR_BG_GANON_OTYUKA", "BG_GANON_OTYUKA", "BG_GANON_OTYUKA"),
-    ("ACTOR_BG_SPOT15_RRBOX", "BG_SPOT15_RRBOX", "BG_SPOT15_RRBOX"),
-    ("ACTOR_BG_UMAJUMP", "BG_UMAJUMP", "BG_UMAJUMP"),
-    ("ACTOR_ARROW_FIRE", "ARROW_FIRE", "ARROW_FIRE"),
-    ("ACTOR_ARROW_ICE", "ARROW_ICE", "ARROW_ICE"),
-    ("ACTOR_ARROW_LIGHT", "ARROW_LIGHT", "ARROW_LIGHT"),
-    ("ACTOR_ITEM_ETCETERA", "ITEM_ETCETERA", "ITEM_ETCETERA"),
-    ("ACTOR_OBJ_KIBAKO", "OBJ_KIBAKO", "OBJ_KIBAKO"),
-    ("ACTOR_OBJ_TSUBO", "OBJ_TSUBO", "OBJ_TSUBO"),
-    ("ACTOR_EN_WONDER_ITEM", "EN_WONDER_ITEM", "EN_WONDER_ITEM"),
-    ("ACTOR_EN_IK", "EN_IK", "EN_IK"),
-    ("ACTOR_DEMO_IK", "DEMO_IK", "DEMO_IK"),
-    ("ACTOR_EN_SKJ", "EN_SKJ", "EN_SKJ"),
-    ("ACTOR_EN_SKJNEEDLE", "EN_SKJNEEDLE", "EN_SKJNEEDLE"),
-    ("ACTOR_EN_G_SWITCH", "EN_G_SWITCH", "EN_G_SWITCH"),
-    ("ACTOR_DEMO_EXT", "DEMO_EXT", "DEMO_EXT"),
-    ("ACTOR_DEMO_SHD", "DEMO_SHD", "DEMO_SHD"),
-    ("ACTOR_EN_DNS", "EN_DNS", "EN_DNS"),
-    ("ACTOR_ELF_MSG", "ELF_MSG", "ELF_MSG"),
-    ("ACTOR_EN_HONOTRAP", "EN_HONOTRAP", "EN_HONOTRAP"),
-    ("ACTOR_EN_TUBO_TRAP", "EN_TUBO_TRAP", "EN_TUBO_TRAP"),
-    ("ACTOR_OBJ_ICE_POLY", "OBJ_ICE_POLY", "OBJ_ICE_POLY"),
-    ("ACTOR_BG_SPOT03_TAKI", "BG_SPOT03_TAKI", "BG_SPOT03_TAKI"),
-    ("ACTOR_BG_SPOT07_TAKI", "BG_SPOT07_TAKI", "BG_SPOT07_TAKI"),
-    ("ACTOR_EN_FZ", "EN_FZ", "EN_FZ"),
-    ("ACTOR_EN_PO_RELAY", "EN_PO_RELAY", "EN_PO_RELAY"),
-    ("ACTOR_BG_RELAY_OBJECTS", "BG_RELAY_OBJECTS", "BG_RELAY_OBJECTS"),
-    ("ACTOR_EN_DIVING_GAME", "EN_DIVING_GAME", "EN_DIVING_GAME"),
-    ("ACTOR_EN_KUSA", "EN_KUSA", "EN_KUSA"),
-    ("ACTOR_OBJ_BEAN", "OBJ_BEAN", "OBJ_BEAN"),
-    ("ACTOR_OBJ_BOMBIWA", "OBJ_BOMBIWA", "OBJ_BOMBIWA"),
-    ("ACTOR_OBJ_SWITCH", "OBJ_SWITCH", "OBJ_SWITCH"),
-    ("ACTOR_OBJ_ELEVATOR", "OBJ_ELEVATOR", "OBJ_ELEVATOR"),
-    ("ACTOR_OBJ_LIFT", "OBJ_LIFT", "OBJ_LIFT"),
-    ("ACTOR_OBJ_HSBLOCK", "OBJ_HSBLOCK", "OBJ_HSBLOCK"),
-    ("ACTOR_EN_OKARINA_TAG", "EN_OKARINA_TAG", "EN_OKARINA_TAG"),
-    ("ACTOR_EN_YABUSAME_MARK", "EN_YABUSAME_MARK", "EN_YABUSAME_MARK"),
-    ("ACTOR_EN_GOROIWA", "EN_GOROIWA", "EN_GOROIWA"),
-    ("ACTOR_EN_EX_RUPPY", "EN_EX_RUPPY", "EN_EX_RUPPY"),
-    ("ACTOR_EN_TORYO", "EN_TORYO", "EN_TORYO"),
-    ("ACTOR_EN_DAIKU", "EN_DAIKU", "EN_DAIKU"),
-    ("ACTOR_EN_NWC", "EN_NWC", "EN_NWC"),
-    ("ACTOR_EN_BLKOBJ", "EN_BLKOBJ", "EN_BLKOBJ"),
-    ("ACTOR_ITEM_INBOX", "ITEM_INBOX", "ITEM_INBOX"),
-    ("ACTOR_EN_GE1", "EN_GE1", "EN_GE1"),
-    ("ACTOR_OBJ_BLOCKSTOP", "OBJ_BLOCKSTOP", "OBJ_BLOCKSTOP"),
-    ("ACTOR_EN_SDA", "EN_SDA", "EN_SDA"),
-    ("ACTOR_EN_CLEAR_TAG", "EN_CLEAR_TAG", "EN_CLEAR_TAG"),
-    ("ACTOR_EN_NIW_LADY", "EN_NIW_LADY", "EN_NIW_LADY"),
-    ("ACTOR_EN_GM", "EN_GM", "EN_GM"),
-    ("ACTOR_EN_MS", "EN_MS", "EN_MS"),
-    ("ACTOR_EN_HS", "EN_HS", "EN_HS"),
-    ("ACTOR_BG_INGATE", "BG_INGATE", "BG_INGATE"),
-    ("ACTOR_EN_KANBAN", "EN_KANBAN", "EN_KANBAN"),
-    ("ACTOR_EN_HEISHI3", "EN_HEISHI3", "EN_HEISHI3"),
-    ("ACTOR_EN_SYATEKI_NIW", "EN_SYATEKI_NIW", "EN_SYATEKI_NIW"),
-    ("ACTOR_EN_ATTACK_NIW", "EN_ATTACK_NIW", "EN_ATTACK_NIW"),
-    ("ACTOR_BG_SPOT01_IDOSOKO", "BG_SPOT01_IDOSOKO", "BG_SPOT01_IDOSOKO"),
-    ("ACTOR_EN_SA", "EN_SA", "EN_SA"),
-    ("ACTOR_EN_WONDER_TALK", "EN_WONDER_TALK", "EN_WONDER_TALK"),
-    ("ACTOR_BG_GJYO_BRIDGE", "BG_GJYO_BRIDGE", "BG_GJYO_BRIDGE"),
-    ("ACTOR_EN_DS", "EN_DS", "EN_DS"),
-    ("ACTOR_EN_MK", "EN_MK", "EN_MK"),
-    ("ACTOR_EN_BOM_BOWL_MAN", "EN_BOM_BOWL_MAN", "EN_BOM_BOWL_MAN"),
-    ("ACTOR_EN_BOM_BOWL_PIT", "EN_BOM_BOWL_PIT", "EN_BOM_BOWL_PIT"),
-    ("ACTOR_EN_OWL", "EN_OWL", "EN_OWL"),
-    ("ACTOR_EN_ISHI", "EN_ISHI", "EN_ISHI"),
-    ("ACTOR_OBJ_HANA", "OBJ_HANA", "OBJ_HANA"),
-    ("ACTOR_OBJ_LIGHTSWITCH", "OBJ_LIGHTSWITCH", "OBJ_LIGHTSWITCH"),
-    ("ACTOR_OBJ_MURE2", "OBJ_MURE2", "OBJ_MURE2"),
-    ("ACTOR_EN_GO", "EN_GO", "EN_GO"),
-    ("ACTOR_EN_FU", "EN_FU", "EN_FU"),
-    ("ACTOR_EN_CHANGER", "EN_CHANGER", "EN_CHANGER"),
-    ("ACTOR_BG_JYA_MEGAMI", "BG_JYA_MEGAMI", "BG_JYA_MEGAMI"),
-    ("ACTOR_BG_JYA_LIFT", "BG_JYA_LIFT", "BG_JYA_LIFT"),
-    ("ACTOR_BG_JYA_BIGMIRROR", "BG_JYA_BIGMIRROR", "BG_JYA_BIGMIRROR"),
-    ("ACTOR_BG_JYA_BOMBCHUIWA", "BG_JYA_BOMBCHUIWA", "BG_JYA_BOMBCHUIWA"),
-    ("ACTOR_BG_JYA_AMISHUTTER", "BG_JYA_AMISHUTTER", "BG_JYA_AMISHUTTER"),
-    ("ACTOR_BG_JYA_BOMBIWA", "BG_JYA_BOMBIWA", "BG_JYA_BOMBIWA"),
-    ("ACTOR_BG_SPOT18_BASKET", "BG_SPOT18_BASKET", "BG_SPOT18_BASKET"),
-    ("ACTOR_EN_GANON_ORGAN", "EN_GANON_ORGAN", "EN_GANON_ORGAN"),
-    ("ACTOR_EN_SIOFUKI", "EN_SIOFUKI", "EN_SIOFUKI"),
-    ("ACTOR_EN_STREAM", "EN_STREAM", "EN_STREAM"),
-    ("ACTOR_EN_MM", "EN_MM", "EN_MM"),
-    ("ACTOR_EN_KO", "EN_KO", "EN_KO"),
-    ("ACTOR_EN_KZ", "EN_KZ", "EN_KZ"),
-    ("ACTOR_EN_WEATHER_TAG", "EN_WEATHER_TAG", "EN_WEATHER_TAG"),
-    ("ACTOR_BG_SST_FLOOR", "BG_SST_FLOOR", "BG_SST_FLOOR"),
-    ("ACTOR_EN_ANI", "EN_ANI", "EN_ANI"),
-    ("ACTOR_EN_EX_ITEM", "EN_EX_ITEM", "EN_EX_ITEM"),
-    ("ACTOR_BG_JYA_IRONOBJ", "BG_JYA_IRONOBJ", "BG_JYA_IRONOBJ"),
-    ("ACTOR_EN_JS", "EN_JS", "EN_JS"),
-    ("ACTOR_EN_JSJUTAN", "EN_JSJUTAN", "EN_JSJUTAN"),
-    ("ACTOR_EN_CS", "EN_CS", "EN_CS"),
-    ("ACTOR_EN_MD", "EN_MD", "EN_MD"),
-    ("ACTOR_EN_HY", "EN_HY", "EN_HY"),
-    ("ACTOR_EN_GANON_MANT", "EN_GANON_MANT", "EN_GANON_MANT"),
-    ("ACTOR_EN_OKARINA_EFFECT", "EN_OKARINA_EFFECT", "EN_OKARINA_EFFECT"),
-    ("ACTOR_EN_MAG", "EN_MAG", "EN_MAG"),
-    ("ACTOR_DOOR_GERUDO", "DOOR_GERUDO", "DOOR_GERUDO"),
-    ("ACTOR_ELF_MSG2", "ELF_MSG2", "ELF_MSG2"),
-    ("ACTOR_DEMO_GT", "DEMO_GT", "DEMO_GT"),
-    ("ACTOR_EN_PO_FIELD", "EN_PO_FIELD", "EN_PO_FIELD"),
-    ("ACTOR_EFC_ERUPC", "EFC_ERUPC", "EFC_ERUPC"),
-    ("ACTOR_BG_ZG", "BG_ZG", "BG_ZG"),
-    ("ACTOR_EN_HEISHI4", "EN_HEISHI4", "EN_HEISHI4"),
-    ("ACTOR_EN_ZL3", "EN_ZL3", "EN_ZL3"),
-    ("ACTOR_BOSS_GANON2", "BOSS_GANON2", "BOSS_GANON2"),
-    ("ACTOR_EN_KAKASI", "EN_KAKASI", "EN_KAKASI"),
-    ("ACTOR_EN_TAKARA_MAN", "EN_TAKARA_MAN", "EN_TAKARA_MAN"),
-    ("ACTOR_OBJ_MAKEOSHIHIKI", "OBJ_MAKEOSHIHIKI", "OBJ_MAKEOSHIHIKI"),
-    ("ACTOR_OCEFF_SPOT", "OCEFF_SPOT", "OCEFF_SPOT"),
-    ("ACTOR_END_TITLE", "END_TITLE", "END_TITLE"),
-    ("ACTOR_EN_TORCH", "EN_TORCH", "EN_TORCH"),
-    ("ACTOR_DEMO_EC", "DEMO_EC", "DEMO_EC"),
-    ("ACTOR_SHOT_SUN", "SHOT_SUN", "SHOT_SUN"),
-    ("ACTOR_EN_DY_EXTRA", "EN_DY_EXTRA", "EN_DY_EXTRA"),
-    ("ACTOR_EN_WONDER_TALK2", "EN_WONDER_TALK2", "EN_WONDER_TALK2"),
-    ("ACTOR_EN_GE2", "EN_GE2", "EN_GE2"),
-    ("ACTOR_OBJ_ROOMTIMER", "OBJ_ROOMTIMER", "OBJ_ROOMTIMER"),
-    ("ACTOR_EN_SSH", "EN_SSH", "EN_SSH"),
-    ("ACTOR_EN_STH", "EN_STH", "EN_STH"),
-    ("ACTOR_OCEFF_WIPE", "OCEFF_WIPE", "OCEFF_WIPE"),
-    ("ACTOR_OCEFF_STORM", "OCEFF_STORM", "OCEFF_STORM"),
-    ("ACTOR_EN_WEIYER", "EN_WEIYER", "EN_WEIYER"),
-    ("ACTOR_BG_SPOT05_SOKO", "BG_SPOT05_SOKO", "BG_SPOT05_SOKO"),
-    ("ACTOR_BG_JYA_1FLIFT", "BG_JYA_1FLIFT", "BG_JYA_1FLIFT"),
-    ("ACTOR_BG_JYA_HAHENIRON", "BG_JYA_HAHENIRON", "BG_JYA_HAHENIRON"),
-    ("ACTOR_BG_SPOT12_GATE", "BG_SPOT12_GATE", "BG_SPOT12_GATE"),
-    ("ACTOR_BG_SPOT12_SAKU", "BG_SPOT12_SAKU", "BG_SPOT12_SAKU"),
-    ("ACTOR_EN_HINTNUTS", "EN_HINTNUTS", "EN_HINTNUTS"),
-    ("ACTOR_EN_NUTSBALL", "EN_NUTSBALL", "EN_NUTSBALL"),
-    ("ACTOR_BG_SPOT00_BREAK", "BG_SPOT00_BREAK", "BG_SPOT00_BREAK"),
-    ("ACTOR_EN_SHOPNUTS", "EN_SHOPNUTS", "EN_SHOPNUTS"),
-    ("ACTOR_EN_IT", "EN_IT", "EN_IT"),
-    ("ACTOR_EN_GELDB", "EN_GELDB", "EN_GELDB"),
-    ("ACTOR_OCEFF_WIPE2", "OCEFF_WIPE2", "OCEFF_WIPE2"),
-    ("ACTOR_OCEFF_WIPE3", "OCEFF_WIPE3", "OCEFF_WIPE3"),
-    ("ACTOR_EN_NIW_GIRL", "EN_NIW_GIRL", "EN_NIW_GIRL"),
-    ("ACTOR_EN_DOG", "EN_DOG", "EN_DOG"),
-    ("ACTOR_EN_SI", "EN_SI", "EN_SI"),
-    ("ACTOR_BG_SPOT01_OBJECTS2", "BG_SPOT01_OBJECTS2", "BG_SPOT01_OBJECTS2"),
-    ("ACTOR_OBJ_COMB", "OBJ_COMB", "OBJ_COMB"),
-    ("ACTOR_BG_SPOT11_BAKUDANKABE", "BG_SPOT11_BAKUDANKABE", "BG_SPOT11_BAKUDANKABE"),
-    ("ACTOR_OBJ_KIBAKO2", "OBJ_KIBAKO2", "OBJ_KIBAKO2"),
-    ("ACTOR_EN_DNT_DEMO", "EN_DNT_DEMO", "EN_DNT_DEMO"),
-    ("ACTOR_EN_DNT_JIJI", "EN_DNT_JIJI", "EN_DNT_JIJI"),
-    ("ACTOR_EN_DNT_NOMAL", "EN_DNT_NOMAL", "EN_DNT_NOMAL"),
-    ("ACTOR_EN_GUEST", "EN_GUEST", "EN_GUEST"),
-    ("ACTOR_BG_BOM_GUARD", "BG_BOM_GUARD", "BG_BOM_GUARD"),
-    ("ACTOR_EN_HS2", "EN_HS2", "EN_HS2"),
-    ("ACTOR_DEMO_KEKKAI", "DEMO_KEKKAI", "DEMO_KEKKAI"),
-    ("ACTOR_BG_SPOT08_BAKUDANKABE", "BG_SPOT08_BAKUDANKABE", "BG_SPOT08_BAKUDANKABE"),
-    ("ACTOR_BG_SPOT17_BAKUDANKABE", "BG_SPOT17_BAKUDANKABE", "BG_SPOT17_BAKUDANKABE"),
-    ("ACTOR_OBJ_MURE3", "OBJ_MURE3", "OBJ_MURE3"),
-    ("ACTOR_EN_TG", "EN_TG", "EN_TG"),
-    ("ACTOR_EN_MU", "EN_MU", "EN_MU"),
-    ("ACTOR_EN_GO2", "EN_GO2", "EN_GO2"),
-    ("ACTOR_EN_WF", "EN_WF", "EN_WF"),
-    ("ACTOR_EN_SKB", "EN_SKB", "EN_SKB"),
-    ("ACTOR_DEMO_GJ", "DEMO_GJ", "DEMO_GJ"),
-    ("ACTOR_DEMO_GEFF", "DEMO_GEFF", "DEMO_GEFF"),
-    ("ACTOR_BG_GND_FIREMEIRO", "BG_GND_FIREMEIRO", "BG_GND_FIREMEIRO"),
-    ("ACTOR_BG_GND_DARKMEIRO", "BG_GND_DARKMEIRO", "BG_GND_DARKMEIRO"),
-    ("ACTOR_BG_GND_SOULMEIRO", "BG_GND_SOULMEIRO", "BG_GND_SOULMEIRO"),
-    ("ACTOR_BG_GND_NISEKABE", "BG_GND_NISEKABE", "BG_GND_NISEKABE"),
-    ("ACTOR_BG_GND_ICEBLOCK", "BG_GND_ICEBLOCK", "BG_GND_ICEBLOCK"),
-    ("ACTOR_EN_GB", "EN_GB", "EN_GB"),
-    ("ACTOR_EN_GS", "EN_GS", "EN_GS"),
-    ("ACTOR_BG_MIZU_BWALL", "BG_MIZU_BWALL", "BG_MIZU_BWALL"),
-    ("ACTOR_BG_MIZU_SHUTTER", "BG_MIZU_SHUTTER", "BG_MIZU_SHUTTER"),
-    ("ACTOR_EN_DAIKU_KAKARIKO", "EN_DAIKU_KAKARIKO", "EN_DAIKU_KAKARIKO"),
-    ("ACTOR_BG_BOWL_WALL", "BG_BOWL_WALL", "BG_BOWL_WALL"),
-    ("ACTOR_EN_WALL_TUBO", "EN_WALL_TUBO", "EN_WALL_TUBO"),
-    ("ACTOR_EN_PO_DESERT", "EN_PO_DESERT", "EN_PO_DESERT"),
-    ("ACTOR_EN_CROW", "EN_CROW", "EN_CROW"),
-    ("ACTOR_DOOR_KILLER", "DOOR_KILLER", "DOOR_KILLER"),
-    ("ACTOR_BG_SPOT11_OASIS", "BG_SPOT11_OASIS", "BG_SPOT11_OASIS"),
-    ("ACTOR_BG_SPOT18_FUTA", "BG_SPOT18_FUTA", "BG_SPOT18_FUTA"),
-    ("ACTOR_BG_SPOT18_SHUTTER", "BG_SPOT18_SHUTTER", "BG_SPOT18_SHUTTER"),
-    ("ACTOR_EN_MA3", "EN_MA3", "EN_MA3"),
-    ("ACTOR_EN_COW", "EN_COW", "EN_COW"),
-    ("ACTOR_BG_ICE_TURARA", "BG_ICE_TURARA", "BG_ICE_TURARA"),
-    ("ACTOR_BG_ICE_SHUTTER", "BG_ICE_SHUTTER", "BG_ICE_SHUTTER"),
-    ("ACTOR_EN_KAKASI2", "EN_KAKASI2", "EN_KAKASI2"),
-    ("ACTOR_EN_KAKASI3", "EN_KAKASI3", "EN_KAKASI3"),
-    ("ACTOR_OCEFF_WIPE4", "OCEFF_WIPE4", "OCEFF_WIPE4"),
-    ("ACTOR_EN_EG", "EN_EG", "EN_EG"),
-    ("ACTOR_BG_MENKURI_NISEKABE", "BG_MENKURI_NISEKABE", "BG_MENKURI_NISEKABE"),
-    ("ACTOR_EN_ZO", "EN_ZO", "EN_ZO"),
-    ("ACTOR_OBJ_MAKEKINSUTA", "OBJ_MAKEKINSUTA", "OBJ_MAKEKINSUTA"),
-    ("ACTOR_EN_GE3", "EN_GE3", "EN_GE3"),
-    ("ACTOR_OBJ_TIMEBLOCK", "OBJ_TIMEBLOCK", "OBJ_TIMEBLOCK"),
-    ("ACTOR_OBJ_HAMISHI", "OBJ_HAMISHI", "OBJ_HAMISHI"),
-    ("ACTOR_EN_ZL4", "EN_ZL4", "EN_ZL4"),
-    ("ACTOR_EN_MM2", "EN_MM2", "EN_MM2"),
-    ("ACTOR_BG_JYA_BLOCK", "BG_JYA_BLOCK", "BG_JYA_BLOCK"),
-    ("ACTOR_OBJ_WARP2BLOCK", "OBJ_WARP2BLOCK", "OBJ_WARP2BLOCK"),
+	("ACTOR_EN_TEST", "EN_TEST", "EN_TEST"),
+	("ACTOR_EN_GIRLA", "EN_GIRLA", "EN_GIRLA"),
+	("ACTOR_EN_PART", "EN_PART", "EN_PART"),
+	("ACTOR_EN_LIGHT", "EN_LIGHT", "EN_LIGHT"),
+	("ACTOR_EN_DOOR", "EN_DOOR", "EN_DOOR"),
+	("ACTOR_EN_BOX", "EN_BOX", "EN_BOX"),
+	("ACTOR_BG_DY_YOSEIZO", "BG_DY_YOSEIZO", "BG_DY_YOSEIZO"),
+	("ACTOR_BG_HIDAN_FIREWALL", "BG_HIDAN_FIREWALL", "BG_HIDAN_FIREWALL"),
+	("ACTOR_EN_POH", "EN_POH", "EN_POH"),
+	("ACTOR_EN_OKUTA", "EN_OKUTA", "EN_OKUTA"),
+	("ACTOR_BG_YDAN_SP", "BG_YDAN_SP", "BG_YDAN_SP"),
+	("ACTOR_EN_BOM", "EN_BOM", "EN_BOM"),
+	("ACTOR_EN_WALLMAS", "EN_WALLMAS", "EN_WALLMAS"),
+	("ACTOR_EN_DODONGO", "EN_DODONGO", "EN_DODONGO"),
+	("ACTOR_EN_FIREFLY", "EN_FIREFLY", "EN_FIREFLY"),
+	("ACTOR_EN_HORSE", "EN_HORSE", "EN_HORSE"),
+	("ACTOR_EN_ITEM00", "EN_ITEM00", "EN_ITEM00"),
+	("ACTOR_EN_ARROW", "EN_ARROW", "EN_ARROW"),
+	("ACTOR_EN_ELF", "EN_ELF", "EN_ELF"),
+	("ACTOR_EN_NIW", "EN_NIW", "EN_NIW"),
+	("ACTOR_EN_TITE", "EN_TITE", "EN_TITE"),
+	("ACTOR_EN_REEBA", "EN_REEBA", "EN_REEBA"),
+	("ACTOR_EN_PEEHAT", "EN_PEEHAT", "EN_PEEHAT"),
+	("ACTOR_EN_BUTTE", "EN_BUTTE", "EN_BUTTE"),
+	("ACTOR_EN_INSECT", "EN_INSECT", "EN_INSECT"),
+	("ACTOR_EN_FISH", "EN_FISH", "EN_FISH"),
+	("ACTOR_EN_HOLL", "EN_HOLL", "EN_HOLL"),
+	("ACTOR_EN_SCENE_CHANGE", "EN_SCENE_CHANGE", "EN_SCENE_CHANGE"),
+	("ACTOR_EN_ZF", "EN_ZF", "EN_ZF"),
+	("ACTOR_EN_HATA", "EN_HATA", "EN_HATA"),
+	("ACTOR_BOSS_DODONGO", "BOSS_DODONGO", "BOSS_DODONGO"),
+	("ACTOR_BOSS_GOMA", "BOSS_GOMA", "BOSS_GOMA"),
+	("ACTOR_EN_ZL1", "EN_ZL1", "EN_ZL1"),
+	("ACTOR_EN_VIEWER", "EN_VIEWER", "EN_VIEWER"),
+	("ACTOR_EN_GOMA", "EN_GOMA", "EN_GOMA"),
+	("ACTOR_BG_PUSHBOX", "BG_PUSHBOX", "BG_PUSHBOX"),
+	("ACTOR_EN_BUBBLE", "EN_BUBBLE", "EN_BUBBLE"),
+	("ACTOR_DOOR_SHUTTER", "DOOR_SHUTTER", "DOOR_SHUTTER"),
+	("ACTOR_EN_DODOJR", "EN_DODOJR", "EN_DODOJR"),
+	("ACTOR_EN_BDFIRE", "EN_BDFIRE", "EN_BDFIRE"),
+	("ACTOR_EN_BOOM", "EN_BOOM", "EN_BOOM"),
+	("ACTOR_EN_TORCH2", "EN_TORCH2", "EN_TORCH2"),
+	("ACTOR_EN_BILI", "EN_BILI", "EN_BILI"),
+	("ACTOR_EN_TP", "EN_TP", "EN_TP"),
+	("ACTOR_EN_ST", "EN_ST", "EN_ST"),
+	("ACTOR_EN_BW", "EN_BW", "EN_BW"),
+	("ACTOR_EN_A_OBJ", "EN_A_OBJ", "EN_A_OBJ"),
+	("ACTOR_EN_EIYER", "EN_EIYER", "EN_EIYER"),
+	("ACTOR_EN_RIVER_SOUND", "EN_RIVER_SOUND", "EN_RIVER_SOUND"),
+	("ACTOR_EN_HORSE_NORMAL", "EN_HORSE_NORMAL", "EN_HORSE_NORMAL"),
+	("ACTOR_EN_OSSAN", "EN_OSSAN", "EN_OSSAN"),
+	("ACTOR_BG_TREEMOUTH", "BG_TREEMOUTH", "BG_TREEMOUTH"),
+	("ACTOR_BG_DODOAGO", "BG_DODOAGO", "BG_DODOAGO"),
+	("ACTOR_BG_HIDAN_DALM", "BG_HIDAN_DALM", "BG_HIDAN_DALM"),
+	("ACTOR_BG_HIDAN_HROCK", "BG_HIDAN_HROCK", "BG_HIDAN_HROCK"),
+	("ACTOR_EN_HORSE_GANON", "EN_HORSE_GANON", "EN_HORSE_GANON"),
+	("ACTOR_BG_HIDAN_ROCK", "BG_HIDAN_ROCK", "BG_HIDAN_ROCK"),
+	("ACTOR_BG_HIDAN_RSEKIZOU", "BG_HIDAN_RSEKIZOU", "BG_HIDAN_RSEKIZOU"),
+	("ACTOR_BG_HIDAN_SEKIZOU", "BG_HIDAN_SEKIZOU", "BG_HIDAN_SEKIZOU"),
+	("ACTOR_BG_HIDAN_SIMA", "BG_HIDAN_SIMA", "BG_HIDAN_SIMA"),
+	("ACTOR_BG_HIDAN_SYOKU", "BG_HIDAN_SYOKU", "BG_HIDAN_SYOKU"),
+	("ACTOR_EN_XC", "EN_XC", "EN_XC"),
+	("ACTOR_BG_HIDAN_CURTAIN", "BG_HIDAN_CURTAIN", "BG_HIDAN_CURTAIN"),
+	("ACTOR_BG_SPOT00_HANEBASI", "BG_SPOT00_HANEBASI", "BG_SPOT00_HANEBASI"),
+	("ACTOR_EN_MB", "EN_MB", "EN_MB"),
+	("ACTOR_EN_BOMBF", "EN_BOMBF", "EN_BOMBF"),
+	("ACTOR_EN_ZL2", "EN_ZL2", "EN_ZL2"),
+	("ACTOR_BG_HIDAN_FSLIFT", "BG_HIDAN_FSLIFT", "BG_HIDAN_FSLIFT"),
+	("ACTOR_EN_OE2", "EN_OE2", "EN_OE2"),
+	("ACTOR_BG_YDAN_HASI", "BG_YDAN_HASI", "BG_YDAN_HASI"),
+	("ACTOR_BG_YDAN_MARUTA", "BG_YDAN_MARUTA", "BG_YDAN_MARUTA"),
+	("ACTOR_BOSS_GANONDROF", "BOSS_GANONDROF", "BOSS_GANONDROF"),
+	("ACTOR_EN_AM", "EN_AM", "EN_AM"),
+	("ACTOR_EN_DEKUBABA", "EN_DEKUBABA", "EN_DEKUBABA"),
+	("ACTOR_EN_M_FIRE1", "EN_M_FIRE1", "EN_M_FIRE1"),
+	("ACTOR_EN_M_THUNDER", "EN_M_THUNDER", "EN_M_THUNDER"),
+	("ACTOR_BG_DDAN_JD", "BG_DDAN_JD", "BG_DDAN_JD"),
+	("ACTOR_BG_BREAKWALL", "BG_BREAKWALL", "BG_BREAKWALL"),
+	("ACTOR_EN_JJ", "EN_JJ", "EN_JJ"),
+	("ACTOR_EN_HORSE_ZELDA", "EN_HORSE_ZELDA", "EN_HORSE_ZELDA"),
+	("ACTOR_BG_DDAN_KD", "BG_DDAN_KD", "BG_DDAN_KD"),
+	("ACTOR_DOOR_WARP1", "DOOR_WARP1", "DOOR_WARP1"),
+	("ACTOR_OBJ_SYOKUDAI", "OBJ_SYOKUDAI", "OBJ_SYOKUDAI"),
+	("ACTOR_ITEM_B_HEART", "ITEM_B_HEART", "ITEM_B_HEART"),
+	("ACTOR_EN_DEKUNUTS", "EN_DEKUNUTS", "EN_DEKUNUTS"),
+	("ACTOR_BG_MENKURI_KAITEN", "BG_MENKURI_KAITEN", "BG_MENKURI_KAITEN"),
+	("ACTOR_BG_MENKURI_EYE", "BG_MENKURI_EYE", "BG_MENKURI_EYE"),
+	("ACTOR_EN_VALI", "EN_VALI", "EN_VALI"),
+	("ACTOR_BG_MIZU_MOVEBG", "BG_MIZU_MOVEBG", "BG_MIZU_MOVEBG"),
+	("ACTOR_BG_MIZU_WATER", "BG_MIZU_WATER", "BG_MIZU_WATER"),
+	("ACTOR_ARMS_HOOK", "ARMS_HOOK", "ARMS_HOOK"),
+	("ACTOR_EN_FHG", "EN_FHG", "EN_FHG"),
+	("ACTOR_BG_MORI_HINERI", "BG_MORI_HINERI", "BG_MORI_HINERI"),
+	("ACTOR_EN_BB", "EN_BB", "EN_BB"),
+	("ACTOR_BG_TOKI_HIKARI", "BG_TOKI_HIKARI", "BG_TOKI_HIKARI"),
+	("ACTOR_EN_YUKABYUN", "EN_YUKABYUN", "EN_YUKABYUN"),
+	("ACTOR_BG_TOKI_SWD", "BG_TOKI_SWD", "BG_TOKI_SWD"),
+	("ACTOR_EN_FHG_FIRE", "EN_FHG_FIRE", "EN_FHG_FIRE"),
+	("ACTOR_BG_MJIN", "BG_MJIN", "BG_MJIN"),
+	("ACTOR_BG_HIDAN_KOUSI", "BG_HIDAN_KOUSI", "BG_HIDAN_KOUSI"),
+	("ACTOR_DOOR_TOKI", "DOOR_TOKI", "DOOR_TOKI"),
+	("ACTOR_BG_HIDAN_HAMSTEP", "BG_HIDAN_HAMSTEP", "BG_HIDAN_HAMSTEP"),
+	("ACTOR_EN_BIRD", "EN_BIRD", "EN_BIRD"),
+	("ACTOR_EN_WOOD02", "EN_WOOD02", "EN_WOOD02"),
+	("ACTOR_EN_LIGHTBOX", "EN_LIGHTBOX", "EN_LIGHTBOX"),
+	("ACTOR_EN_PU_BOX", "EN_PU_BOX", "EN_PU_BOX"),
+	("ACTOR_EN_TRAP", "EN_TRAP", "EN_TRAP"),
+	("ACTOR_EN_AROW_TRAP", "EN_AROW_TRAP", "EN_AROW_TRAP"),
+	("ACTOR_EN_VASE", "EN_VASE", "EN_VASE"),
+	("ACTOR_EN_TA", "EN_TA", "EN_TA"),
+	("ACTOR_EN_TK", "EN_TK", "EN_TK"),
+	("ACTOR_BG_MORI_BIGST", "BG_MORI_BIGST", "BG_MORI_BIGST"),
+	("ACTOR_BG_MORI_ELEVATOR", "BG_MORI_ELEVATOR", "BG_MORI_ELEVATOR"),
+	("ACTOR_BG_MORI_KAITENKABE", "BG_MORI_KAITENKABE", "BG_MORI_KAITENKABE"),
+	("ACTOR_BG_MORI_RAKKATENJO", "BG_MORI_RAKKATENJO", "BG_MORI_RAKKATENJO"),
+	("ACTOR_EN_VM", "EN_VM", "EN_VM"),
+	("ACTOR_DEMO_EFFECT", "DEMO_EFFECT", "DEMO_EFFECT"),
+	("ACTOR_DEMO_KANKYO", "DEMO_KANKYO", "DEMO_KANKYO"),
+	("ACTOR_BG_HIDAN_FWBIG", "BG_HIDAN_FWBIG", "BG_HIDAN_FWBIG"),
+	("ACTOR_EN_FLOORMAS", "EN_FLOORMAS", "EN_FLOORMAS"),
+	("ACTOR_EN_HEISHI1", "EN_HEISHI1", "EN_HEISHI1"),
+	("ACTOR_EN_RD", "EN_RD", "EN_RD"),
+	("ACTOR_EN_PO_SISTERS", "EN_PO_SISTERS", "EN_PO_SISTERS"),
+	("ACTOR_BG_HEAVY_BLOCK", "BG_HEAVY_BLOCK", "BG_HEAVY_BLOCK"),
+	("ACTOR_BG_PO_EVENT", "BG_PO_EVENT", "BG_PO_EVENT"),
+	("ACTOR_OBJ_MURE", "OBJ_MURE", "OBJ_MURE"),
+	("ACTOR_EN_SW", "EN_SW", "EN_SW"),
+	("ACTOR_BOSS_FD", "BOSS_FD", "BOSS_FD"),
+	("ACTOR_OBJECT_KANKYO", "OBJECT_KANKYO", "OBJECT_KANKYO"),
+	("ACTOR_EN_DU", "EN_DU", "EN_DU"),
+	("ACTOR_EN_FD", "EN_FD", "EN_FD"),
+	("ACTOR_EN_HORSE_LINK_CHILD", "EN_HORSE_LINK_CHILD", "EN_HORSE_LINK_CHILD"),
+	("ACTOR_DOOR_ANA", "DOOR_ANA", "DOOR_ANA"),
+	("ACTOR_BG_SPOT02_OBJECTS", "BG_SPOT02_OBJECTS", "BG_SPOT02_OBJECTS"),
+	("ACTOR_BG_HAKA", "BG_HAKA", "BG_HAKA"),
+	("ACTOR_MAGIC_WIND", "MAGIC_WIND", "MAGIC_WIND"),
+	("ACTOR_MAGIC_FIRE", "MAGIC_FIRE", "MAGIC_FIRE"),
+	("ACTOR_EN_RU1", "EN_RU1", "EN_RU1"),
+	("ACTOR_BOSS_FD2", "BOSS_FD2", "BOSS_FD2"),
+	("ACTOR_EN_FD_FIRE", "EN_FD_FIRE", "EN_FD_FIRE"),
+	("ACTOR_EN_DH", "EN_DH", "EN_DH"),
+	("ACTOR_EN_DHA", "EN_DHA", "EN_DHA"),
+	("ACTOR_EN_RL", "EN_RL", "EN_RL"),
+	("ACTOR_EN_ENCOUNT1", "EN_ENCOUNT1", "EN_ENCOUNT1"),
+	("ACTOR_DEMO_DU", "DEMO_DU", "DEMO_DU"),
+	("ACTOR_DEMO_IM", "DEMO_IM", "DEMO_IM"),
+	("ACTOR_DEMO_TRE_LGT", "DEMO_TRE_LGT", "DEMO_TRE_LGT"),
+	("ACTOR_EN_FW", "EN_FW", "EN_FW"),
+	("ACTOR_BG_VB_SIMA", "BG_VB_SIMA", "BG_VB_SIMA"),
+	("ACTOR_EN_VB_BALL", "EN_VB_BALL", "EN_VB_BALL"),
+	("ACTOR_BG_HAKA_MEGANE", "BG_HAKA_MEGANE", "BG_HAKA_MEGANE"),
+	("ACTOR_BG_HAKA_MEGANEBG", "BG_HAKA_MEGANEBG", "BG_HAKA_MEGANEBG"),
+	("ACTOR_BG_HAKA_SHIP", "BG_HAKA_SHIP", "BG_HAKA_SHIP"),
+	("ACTOR_BG_HAKA_SGAMI", "BG_HAKA_SGAMI", "BG_HAKA_SGAMI"),
+	("ACTOR_EN_HEISHI2", "EN_HEISHI2", "EN_HEISHI2"),
+	("ACTOR_EN_ENCOUNT2", "EN_ENCOUNT2", "EN_ENCOUNT2"),
+	("ACTOR_EN_FIRE_ROCK", "EN_FIRE_ROCK", "EN_FIRE_ROCK"),
+	("ACTOR_EN_BROB", "EN_BROB", "EN_BROB"),
+	("ACTOR_MIR_RAY", "MIR_RAY", "MIR_RAY"),
+	("ACTOR_BG_SPOT09_OBJ", "BG_SPOT09_OBJ", "BG_SPOT09_OBJ"),
+	("ACTOR_BG_SPOT18_OBJ", "BG_SPOT18_OBJ", "BG_SPOT18_OBJ"),
+	("ACTOR_BOSS_VA", "BOSS_VA", "BOSS_VA"),
+	("ACTOR_BG_HAKA_TUBO", "BG_HAKA_TUBO", "BG_HAKA_TUBO"),
+	("ACTOR_BG_HAKA_TRAP", "BG_HAKA_TRAP", "BG_HAKA_TRAP"),
+	("ACTOR_BG_HAKA_HUTA", "BG_HAKA_HUTA", "BG_HAKA_HUTA"),
+	("ACTOR_BG_HAKA_ZOU", "BG_HAKA_ZOU", "BG_HAKA_ZOU"),
+	("ACTOR_BG_SPOT17_FUNEN", "BG_SPOT17_FUNEN", "BG_SPOT17_FUNEN"),
+	("ACTOR_EN_SYATEKI_ITM", "EN_SYATEKI_ITM", "EN_SYATEKI_ITM"),
+	("ACTOR_EN_SYATEKI_MAN", "EN_SYATEKI_MAN", "EN_SYATEKI_MAN"),
+	("ACTOR_EN_TANA", "EN_TANA", "EN_TANA"),
+	("ACTOR_EN_NB", "EN_NB", "EN_NB"),
+	("ACTOR_BOSS_MO", "BOSS_MO", "BOSS_MO"),
+	("ACTOR_EN_SB", "EN_SB", "EN_SB"),
+	("ACTOR_EN_BIGOKUTA", "EN_BIGOKUTA", "EN_BIGOKUTA"),
+	("ACTOR_EN_KAREBABA", "EN_KAREBABA", "EN_KAREBABA"),
+	("ACTOR_BG_BDAN_OBJECTS", "BG_BDAN_OBJECTS", "BG_BDAN_OBJECTS"),
+	("ACTOR_DEMO_SA", "DEMO_SA", "DEMO_SA"),
+	("ACTOR_DEMO_GO", "DEMO_GO", "DEMO_GO"),
+	("ACTOR_EN_IN", "EN_IN", "EN_IN"),
+	("ACTOR_EN_TR", "EN_TR", "EN_TR"),
+	("ACTOR_BG_SPOT16_BOMBSTONE", "BG_SPOT16_BOMBSTONE", "BG_SPOT16_BOMBSTONE"),
+	("ACTOR_BG_HIDAN_KOWARERUKABE", "BG_HIDAN_KOWARERUKABE", "BG_HIDAN_KOWARERUKABE"),
+	("ACTOR_BG_BOMBWALL", "BG_BOMBWALL", "BG_BOMBWALL"),
+	("ACTOR_BG_SPOT08_ICEBLOCK", "BG_SPOT08_ICEBLOCK", "BG_SPOT08_ICEBLOCK"),
+	("ACTOR_EN_RU2", "EN_RU2", "EN_RU2"),
+	("ACTOR_OBJ_DEKUJR", "OBJ_DEKUJR", "OBJ_DEKUJR"),
+	("ACTOR_BG_MIZU_UZU", "BG_MIZU_UZU", "BG_MIZU_UZU"),
+	("ACTOR_BG_SPOT06_OBJECTS", "BG_SPOT06_OBJECTS", "BG_SPOT06_OBJECTS"),
+	("ACTOR_BG_ICE_OBJECTS", "BG_ICE_OBJECTS", "BG_ICE_OBJECTS"),
+	("ACTOR_BG_HAKA_WATER", "BG_HAKA_WATER", "BG_HAKA_WATER"),
+	("ACTOR_EN_MA2", "EN_MA2", "EN_MA2"),
+	("ACTOR_EN_BOM_CHU", "EN_BOM_CHU", "EN_BOM_CHU"),
+	("ACTOR_EN_HORSE_GAME_CHECK", "EN_HORSE_GAME_CHECK", "EN_HORSE_GAME_CHECK"),
+	("ACTOR_BOSS_TW", "BOSS_TW", "BOSS_TW"),
+	("ACTOR_EN_RR", "EN_RR", "EN_RR"),
+	("ACTOR_EN_BA", "EN_BA", "EN_BA"),
+	("ACTOR_EN_BX", "EN_BX", "EN_BX"),
+	("ACTOR_EN_ANUBICE", "EN_ANUBICE", "EN_ANUBICE"),
+	("ACTOR_EN_ANUBICE_FIRE", "EN_ANUBICE_FIRE", "EN_ANUBICE_FIRE"),
+	("ACTOR_BG_MORI_HASHIGO", "BG_MORI_HASHIGO", "BG_MORI_HASHIGO"),
+	("ACTOR_BG_MORI_HASHIRA4", "BG_MORI_HASHIRA4", "BG_MORI_HASHIRA4"),
+	("ACTOR_BG_MORI_IDOMIZU", "BG_MORI_IDOMIZU", "BG_MORI_IDOMIZU"),
+	("ACTOR_BG_SPOT16_DOUGHNUT", "BG_SPOT16_DOUGHNUT", "BG_SPOT16_DOUGHNUT"),
+	("ACTOR_BG_BDAN_SWITCH", "BG_BDAN_SWITCH", "BG_BDAN_SWITCH"),
+	("ACTOR_EN_MA1", "EN_MA1", "EN_MA1"),
+	("ACTOR_BOSS_GANON", "BOSS_GANON", "BOSS_GANON"),
+	("ACTOR_BOSS_SST", "BOSS_SST", "BOSS_SST"),
+	("ACTOR_EN_NY", "EN_NY", "EN_NY"),
+	("ACTOR_EN_FR", "EN_FR", "EN_FR"),
+	("ACTOR_ITEM_SHIELD", "ITEM_SHIELD", "ITEM_SHIELD"),
+	("ACTOR_BG_ICE_SHELTER", "BG_ICE_SHELTER", "BG_ICE_SHELTER"),
+	("ACTOR_EN_ICE_HONO", "EN_ICE_HONO", "EN_ICE_HONO"),
+	("ACTOR_ITEM_OCARINA", "ITEM_OCARINA", "ITEM_OCARINA"),
+	("ACTOR_MAGIC_DARK", "MAGIC_DARK", "MAGIC_DARK"),
+	("ACTOR_DEMO_6K", "DEMO_6K", "DEMO_6K"),
+	("ACTOR_EN_ANUBICE_TAG", "EN_ANUBICE_TAG", "EN_ANUBICE_TAG"),
+	("ACTOR_BG_HAKA_GATE", "BG_HAKA_GATE", "BG_HAKA_GATE"),
+	("ACTOR_BG_SPOT15_SAKU", "BG_SPOT15_SAKU", "BG_SPOT15_SAKU"),
+	("ACTOR_BG_JYA_GOROIWA", "BG_JYA_GOROIWA", "BG_JYA_GOROIWA"),
+	("ACTOR_BG_JYA_ZURERUKABE", "BG_JYA_ZURERUKABE", "BG_JYA_ZURERUKABE"),
+	("ACTOR_BG_JYA_COBRA", "BG_JYA_COBRA", "BG_JYA_COBRA"),
+	("ACTOR_BG_JYA_KANAAMI", "BG_JYA_KANAAMI", "BG_JYA_KANAAMI"),
+	("ACTOR_FISHING", "FISHING", "FISHING"),
+	("ACTOR_OBJ_OSHIHIKI", "OBJ_OSHIHIKI", "OBJ_OSHIHIKI"),
+	("ACTOR_BG_GATE_SHUTTER", "BG_GATE_SHUTTER", "BG_GATE_SHUTTER"),
+	("ACTOR_EFF_DUST", "EFF_DUST", "EFF_DUST"),
+	("ACTOR_BG_SPOT01_FUSYA", "BG_SPOT01_FUSYA", "BG_SPOT01_FUSYA"),
+	("ACTOR_BG_SPOT01_IDOHASHIRA", "BG_SPOT01_IDOHASHIRA", "BG_SPOT01_IDOHASHIRA"),
+	("ACTOR_BG_SPOT01_IDOMIZU", "BG_SPOT01_IDOMIZU", "BG_SPOT01_IDOMIZU"),
+	("ACTOR_BG_PO_SYOKUDAI", "BG_PO_SYOKUDAI", "BG_PO_SYOKUDAI"),
+	("ACTOR_BG_GANON_OTYUKA", "BG_GANON_OTYUKA", "BG_GANON_OTYUKA"),
+	("ACTOR_BG_SPOT15_RRBOX", "BG_SPOT15_RRBOX", "BG_SPOT15_RRBOX"),
+	("ACTOR_BG_UMAJUMP", "BG_UMAJUMP", "BG_UMAJUMP"),
+	("ACTOR_ARROW_FIRE", "ARROW_FIRE", "ARROW_FIRE"),
+	("ACTOR_ARROW_ICE", "ARROW_ICE", "ARROW_ICE"),
+	("ACTOR_ARROW_LIGHT", "ARROW_LIGHT", "ARROW_LIGHT"),
+	("ACTOR_ITEM_ETCETERA", "ITEM_ETCETERA", "ITEM_ETCETERA"),
+	("ACTOR_OBJ_KIBAKO", "OBJ_KIBAKO", "OBJ_KIBAKO"),
+	("ACTOR_OBJ_TSUBO", "OBJ_TSUBO", "OBJ_TSUBO"),
+	("ACTOR_EN_WONDER_ITEM", "EN_WONDER_ITEM", "EN_WONDER_ITEM"),
+	("ACTOR_EN_IK", "EN_IK", "EN_IK"),
+	("ACTOR_DEMO_IK", "DEMO_IK", "DEMO_IK"),
+	("ACTOR_EN_SKJ", "EN_SKJ", "EN_SKJ"),
+	("ACTOR_EN_SKJNEEDLE", "EN_SKJNEEDLE", "EN_SKJNEEDLE"),
+	("ACTOR_EN_G_SWITCH", "EN_G_SWITCH", "EN_G_SWITCH"),
+	("ACTOR_DEMO_EXT", "DEMO_EXT", "DEMO_EXT"),
+	("ACTOR_DEMO_SHD", "DEMO_SHD", "DEMO_SHD"),
+	("ACTOR_EN_DNS", "EN_DNS", "EN_DNS"),
+	("ACTOR_ELF_MSG", "ELF_MSG", "ELF_MSG"),
+	("ACTOR_EN_HONOTRAP", "EN_HONOTRAP", "EN_HONOTRAP"),
+	("ACTOR_EN_TUBO_TRAP", "EN_TUBO_TRAP", "EN_TUBO_TRAP"),
+	("ACTOR_OBJ_ICE_POLY", "OBJ_ICE_POLY", "OBJ_ICE_POLY"),
+	("ACTOR_BG_SPOT03_TAKI", "BG_SPOT03_TAKI", "BG_SPOT03_TAKI"),
+	("ACTOR_BG_SPOT07_TAKI", "BG_SPOT07_TAKI", "BG_SPOT07_TAKI"),
+	("ACTOR_EN_FZ", "EN_FZ", "EN_FZ"),
+	("ACTOR_EN_PO_RELAY", "EN_PO_RELAY", "EN_PO_RELAY"),
+	("ACTOR_BG_RELAY_OBJECTS", "BG_RELAY_OBJECTS", "BG_RELAY_OBJECTS"),
+	("ACTOR_EN_DIVING_GAME", "EN_DIVING_GAME", "EN_DIVING_GAME"),
+	("ACTOR_EN_KUSA", "EN_KUSA", "EN_KUSA"),
+	("ACTOR_OBJ_BEAN", "OBJ_BEAN", "OBJ_BEAN"),
+	("ACTOR_OBJ_BOMBIWA", "OBJ_BOMBIWA", "OBJ_BOMBIWA"),
+	("ACTOR_OBJ_SWITCH", "OBJ_SWITCH", "OBJ_SWITCH"),
+	("ACTOR_OBJ_ELEVATOR", "OBJ_ELEVATOR", "OBJ_ELEVATOR"),
+	("ACTOR_OBJ_LIFT", "OBJ_LIFT", "OBJ_LIFT"),
+	("ACTOR_OBJ_HSBLOCK", "OBJ_HSBLOCK", "OBJ_HSBLOCK"),
+	("ACTOR_EN_OKARINA_TAG", "EN_OKARINA_TAG", "EN_OKARINA_TAG"),
+	("ACTOR_EN_YABUSAME_MARK", "EN_YABUSAME_MARK", "EN_YABUSAME_MARK"),
+	("ACTOR_EN_GOROIWA", "EN_GOROIWA", "EN_GOROIWA"),
+	("ACTOR_EN_EX_RUPPY", "EN_EX_RUPPY", "EN_EX_RUPPY"),
+	("ACTOR_EN_TORYO", "EN_TORYO", "EN_TORYO"),
+	("ACTOR_EN_DAIKU", "EN_DAIKU", "EN_DAIKU"),
+	("ACTOR_EN_NWC", "EN_NWC", "EN_NWC"),
+	("ACTOR_EN_BLKOBJ", "EN_BLKOBJ", "EN_BLKOBJ"),
+	("ACTOR_ITEM_INBOX", "ITEM_INBOX", "ITEM_INBOX"),
+	("ACTOR_EN_GE1", "EN_GE1", "EN_GE1"),
+	("ACTOR_OBJ_BLOCKSTOP", "OBJ_BLOCKSTOP", "OBJ_BLOCKSTOP"),
+	("ACTOR_EN_SDA", "EN_SDA", "EN_SDA"),
+	("ACTOR_EN_CLEAR_TAG", "EN_CLEAR_TAG", "EN_CLEAR_TAG"),
+	("ACTOR_EN_NIW_LADY", "EN_NIW_LADY", "EN_NIW_LADY"),
+	("ACTOR_EN_GM", "EN_GM", "EN_GM"),
+	("ACTOR_EN_MS", "EN_MS", "EN_MS"),
+	("ACTOR_EN_HS", "EN_HS", "EN_HS"),
+	("ACTOR_BG_INGATE", "BG_INGATE", "BG_INGATE"),
+	("ACTOR_EN_KANBAN", "EN_KANBAN", "EN_KANBAN"),
+	("ACTOR_EN_HEISHI3", "EN_HEISHI3", "EN_HEISHI3"),
+	("ACTOR_EN_SYATEKI_NIW", "EN_SYATEKI_NIW", "EN_SYATEKI_NIW"),
+	("ACTOR_EN_ATTACK_NIW", "EN_ATTACK_NIW", "EN_ATTACK_NIW"),
+	("ACTOR_BG_SPOT01_IDOSOKO", "BG_SPOT01_IDOSOKO", "BG_SPOT01_IDOSOKO"),
+	("ACTOR_EN_SA", "EN_SA", "EN_SA"),
+	("ACTOR_EN_WONDER_TALK", "EN_WONDER_TALK", "EN_WONDER_TALK"),
+	("ACTOR_BG_GJYO_BRIDGE", "BG_GJYO_BRIDGE", "BG_GJYO_BRIDGE"),
+	("ACTOR_EN_DS", "EN_DS", "EN_DS"),
+	("ACTOR_EN_MK", "EN_MK", "EN_MK"),
+	("ACTOR_EN_BOM_BOWL_MAN", "EN_BOM_BOWL_MAN", "EN_BOM_BOWL_MAN"),
+	("ACTOR_EN_BOM_BOWL_PIT", "EN_BOM_BOWL_PIT", "EN_BOM_BOWL_PIT"),
+	("ACTOR_EN_OWL", "EN_OWL", "EN_OWL"),
+	("ACTOR_EN_ISHI", "EN_ISHI", "EN_ISHI"),
+	("ACTOR_OBJ_HANA", "OBJ_HANA", "OBJ_HANA"),
+	("ACTOR_OBJ_LIGHTSWITCH", "OBJ_LIGHTSWITCH", "OBJ_LIGHTSWITCH"),
+	("ACTOR_OBJ_MURE2", "OBJ_MURE2", "OBJ_MURE2"),
+	("ACTOR_EN_GO", "EN_GO", "EN_GO"),
+	("ACTOR_EN_FU", "EN_FU", "EN_FU"),
+	("ACTOR_EN_CHANGER", "EN_CHANGER", "EN_CHANGER"),
+	("ACTOR_BG_JYA_MEGAMI", "BG_JYA_MEGAMI", "BG_JYA_MEGAMI"),
+	("ACTOR_BG_JYA_LIFT", "BG_JYA_LIFT", "BG_JYA_LIFT"),
+	("ACTOR_BG_JYA_BIGMIRROR", "BG_JYA_BIGMIRROR", "BG_JYA_BIGMIRROR"),
+	("ACTOR_BG_JYA_BOMBCHUIWA", "BG_JYA_BOMBCHUIWA", "BG_JYA_BOMBCHUIWA"),
+	("ACTOR_BG_JYA_AMISHUTTER", "BG_JYA_AMISHUTTER", "BG_JYA_AMISHUTTER"),
+	("ACTOR_BG_JYA_BOMBIWA", "BG_JYA_BOMBIWA", "BG_JYA_BOMBIWA"),
+	("ACTOR_BG_SPOT18_BASKET", "BG_SPOT18_BASKET", "BG_SPOT18_BASKET"),
+	("ACTOR_EN_GANON_ORGAN", "EN_GANON_ORGAN", "EN_GANON_ORGAN"),
+	("ACTOR_EN_SIOFUKI", "EN_SIOFUKI", "EN_SIOFUKI"),
+	("ACTOR_EN_STREAM", "EN_STREAM", "EN_STREAM"),
+	("ACTOR_EN_MM", "EN_MM", "EN_MM"),
+	("ACTOR_EN_KO", "EN_KO", "EN_KO"),
+	("ACTOR_EN_KZ", "EN_KZ", "EN_KZ"),
+	("ACTOR_EN_WEATHER_TAG", "EN_WEATHER_TAG", "EN_WEATHER_TAG"),
+	("ACTOR_BG_SST_FLOOR", "BG_SST_FLOOR", "BG_SST_FLOOR"),
+	("ACTOR_EN_ANI", "EN_ANI", "EN_ANI"),
+	("ACTOR_EN_EX_ITEM", "EN_EX_ITEM", "EN_EX_ITEM"),
+	("ACTOR_BG_JYA_IRONOBJ", "BG_JYA_IRONOBJ", "BG_JYA_IRONOBJ"),
+	("ACTOR_EN_JS", "EN_JS", "EN_JS"),
+	("ACTOR_EN_JSJUTAN", "EN_JSJUTAN", "EN_JSJUTAN"),
+	("ACTOR_EN_CS", "EN_CS", "EN_CS"),
+	("ACTOR_EN_MD", "EN_MD", "EN_MD"),
+	("ACTOR_EN_HY", "EN_HY", "EN_HY"),
+	("ACTOR_EN_GANON_MANT", "EN_GANON_MANT", "EN_GANON_MANT"),
+	("ACTOR_EN_OKARINA_EFFECT", "EN_OKARINA_EFFECT", "EN_OKARINA_EFFECT"),
+	("ACTOR_EN_MAG", "EN_MAG", "EN_MAG"),
+	("ACTOR_DOOR_GERUDO", "DOOR_GERUDO", "DOOR_GERUDO"),
+	("ACTOR_ELF_MSG2", "ELF_MSG2", "ELF_MSG2"),
+	("ACTOR_DEMO_GT", "DEMO_GT", "DEMO_GT"),
+	("ACTOR_EN_PO_FIELD", "EN_PO_FIELD", "EN_PO_FIELD"),
+	("ACTOR_EFC_ERUPC", "EFC_ERUPC", "EFC_ERUPC"),
+	("ACTOR_BG_ZG", "BG_ZG", "BG_ZG"),
+	("ACTOR_EN_HEISHI4", "EN_HEISHI4", "EN_HEISHI4"),
+	("ACTOR_EN_ZL3", "EN_ZL3", "EN_ZL3"),
+	("ACTOR_BOSS_GANON2", "BOSS_GANON2", "BOSS_GANON2"),
+	("ACTOR_EN_KAKASI", "EN_KAKASI", "EN_KAKASI"),
+	("ACTOR_EN_TAKARA_MAN", "EN_TAKARA_MAN", "EN_TAKARA_MAN"),
+	("ACTOR_OBJ_MAKEOSHIHIKI", "OBJ_MAKEOSHIHIKI", "OBJ_MAKEOSHIHIKI"),
+	("ACTOR_OCEFF_SPOT", "OCEFF_SPOT", "OCEFF_SPOT"),
+	("ACTOR_END_TITLE", "END_TITLE", "END_TITLE"),
+	("ACTOR_EN_TORCH", "EN_TORCH", "EN_TORCH"),
+	("ACTOR_DEMO_EC", "DEMO_EC", "DEMO_EC"),
+	("ACTOR_SHOT_SUN", "SHOT_SUN", "SHOT_SUN"),
+	("ACTOR_EN_DY_EXTRA", "EN_DY_EXTRA", "EN_DY_EXTRA"),
+	("ACTOR_EN_WONDER_TALK2", "EN_WONDER_TALK2", "EN_WONDER_TALK2"),
+	("ACTOR_EN_GE2", "EN_GE2", "EN_GE2"),
+	("ACTOR_OBJ_ROOMTIMER", "OBJ_ROOMTIMER", "OBJ_ROOMTIMER"),
+	("ACTOR_EN_SSH", "EN_SSH", "EN_SSH"),
+	("ACTOR_EN_STH", "EN_STH", "EN_STH"),
+	("ACTOR_OCEFF_WIPE", "OCEFF_WIPE", "OCEFF_WIPE"),
+	("ACTOR_OCEFF_STORM", "OCEFF_STORM", "OCEFF_STORM"),
+	("ACTOR_EN_WEIYER", "EN_WEIYER", "EN_WEIYER"),
+	("ACTOR_BG_SPOT05_SOKO", "BG_SPOT05_SOKO", "BG_SPOT05_SOKO"),
+	("ACTOR_BG_JYA_1FLIFT", "BG_JYA_1FLIFT", "BG_JYA_1FLIFT"),
+	("ACTOR_BG_JYA_HAHENIRON", "BG_JYA_HAHENIRON", "BG_JYA_HAHENIRON"),
+	("ACTOR_BG_SPOT12_GATE", "BG_SPOT12_GATE", "BG_SPOT12_GATE"),
+	("ACTOR_BG_SPOT12_SAKU", "BG_SPOT12_SAKU", "BG_SPOT12_SAKU"),
+	("ACTOR_EN_HINTNUTS", "EN_HINTNUTS", "EN_HINTNUTS"),
+	("ACTOR_EN_NUTSBALL", "EN_NUTSBALL", "EN_NUTSBALL"),
+	("ACTOR_BG_SPOT00_BREAK", "BG_SPOT00_BREAK", "BG_SPOT00_BREAK"),
+	("ACTOR_EN_SHOPNUTS", "EN_SHOPNUTS", "EN_SHOPNUTS"),
+	("ACTOR_EN_IT", "EN_IT", "EN_IT"),
+	("ACTOR_EN_GELDB", "EN_GELDB", "EN_GELDB"),
+	("ACTOR_OCEFF_WIPE2", "OCEFF_WIPE2", "OCEFF_WIPE2"),
+	("ACTOR_OCEFF_WIPE3", "OCEFF_WIPE3", "OCEFF_WIPE3"),
+	("ACTOR_EN_NIW_GIRL", "EN_NIW_GIRL", "EN_NIW_GIRL"),
+	("ACTOR_EN_DOG", "EN_DOG", "EN_DOG"),
+	("ACTOR_EN_SI", "EN_SI", "EN_SI"),
+	("ACTOR_BG_SPOT01_OBJECTS2", "BG_SPOT01_OBJECTS2", "BG_SPOT01_OBJECTS2"),
+	("ACTOR_OBJ_COMB", "OBJ_COMB", "OBJ_COMB"),
+	("ACTOR_BG_SPOT11_BAKUDANKABE", "BG_SPOT11_BAKUDANKABE", "BG_SPOT11_BAKUDANKABE"),
+	("ACTOR_OBJ_KIBAKO2", "OBJ_KIBAKO2", "OBJ_KIBAKO2"),
+	("ACTOR_EN_DNT_DEMO", "EN_DNT_DEMO", "EN_DNT_DEMO"),
+	("ACTOR_EN_DNT_JIJI", "EN_DNT_JIJI", "EN_DNT_JIJI"),
+	("ACTOR_EN_DNT_NOMAL", "EN_DNT_NOMAL", "EN_DNT_NOMAL"),
+	("ACTOR_EN_GUEST", "EN_GUEST", "EN_GUEST"),
+	("ACTOR_BG_BOM_GUARD", "BG_BOM_GUARD", "BG_BOM_GUARD"),
+	("ACTOR_EN_HS2", "EN_HS2", "EN_HS2"),
+	("ACTOR_DEMO_KEKKAI", "DEMO_KEKKAI", "DEMO_KEKKAI"),
+	("ACTOR_BG_SPOT08_BAKUDANKABE", "BG_SPOT08_BAKUDANKABE", "BG_SPOT08_BAKUDANKABE"),
+	("ACTOR_BG_SPOT17_BAKUDANKABE", "BG_SPOT17_BAKUDANKABE", "BG_SPOT17_BAKUDANKABE"),
+	("ACTOR_OBJ_MURE3", "OBJ_MURE3", "OBJ_MURE3"),
+	("ACTOR_EN_TG", "EN_TG", "EN_TG"),
+	("ACTOR_EN_MU", "EN_MU", "EN_MU"),
+	("ACTOR_EN_GO2", "EN_GO2", "EN_GO2"),
+	("ACTOR_EN_WF", "EN_WF", "EN_WF"),
+	("ACTOR_EN_SKB", "EN_SKB", "EN_SKB"),
+	("ACTOR_DEMO_GJ", "DEMO_GJ", "DEMO_GJ"),
+	("ACTOR_DEMO_GEFF", "DEMO_GEFF", "DEMO_GEFF"),
+	("ACTOR_BG_GND_FIREMEIRO", "BG_GND_FIREMEIRO", "BG_GND_FIREMEIRO"),
+	("ACTOR_BG_GND_DARKMEIRO", "BG_GND_DARKMEIRO", "BG_GND_DARKMEIRO"),
+	("ACTOR_BG_GND_SOULMEIRO", "BG_GND_SOULMEIRO", "BG_GND_SOULMEIRO"),
+	("ACTOR_BG_GND_NISEKABE", "BG_GND_NISEKABE", "BG_GND_NISEKABE"),
+	("ACTOR_BG_GND_ICEBLOCK", "BG_GND_ICEBLOCK", "BG_GND_ICEBLOCK"),
+	("ACTOR_EN_GB", "EN_GB", "EN_GB"),
+	("ACTOR_EN_GS", "EN_GS", "EN_GS"),
+	("ACTOR_BG_MIZU_BWALL", "BG_MIZU_BWALL", "BG_MIZU_BWALL"),
+	("ACTOR_BG_MIZU_SHUTTER", "BG_MIZU_SHUTTER", "BG_MIZU_SHUTTER"),
+	("ACTOR_EN_DAIKU_KAKARIKO", "EN_DAIKU_KAKARIKO", "EN_DAIKU_KAKARIKO"),
+	("ACTOR_BG_BOWL_WALL", "BG_BOWL_WALL", "BG_BOWL_WALL"),
+	("ACTOR_EN_WALL_TUBO", "EN_WALL_TUBO", "EN_WALL_TUBO"),
+	("ACTOR_EN_PO_DESERT", "EN_PO_DESERT", "EN_PO_DESERT"),
+	("ACTOR_EN_CROW", "EN_CROW", "EN_CROW"),
+	("ACTOR_DOOR_KILLER", "DOOR_KILLER", "DOOR_KILLER"),
+	("ACTOR_BG_SPOT11_OASIS", "BG_SPOT11_OASIS", "BG_SPOT11_OASIS"),
+	("ACTOR_BG_SPOT18_FUTA", "BG_SPOT18_FUTA", "BG_SPOT18_FUTA"),
+	("ACTOR_BG_SPOT18_SHUTTER", "BG_SPOT18_SHUTTER", "BG_SPOT18_SHUTTER"),
+	("ACTOR_EN_MA3", "EN_MA3", "EN_MA3"),
+	("ACTOR_EN_COW", "EN_COW", "EN_COW"),
+	("ACTOR_BG_ICE_TURARA", "BG_ICE_TURARA", "BG_ICE_TURARA"),
+	("ACTOR_BG_ICE_SHUTTER", "BG_ICE_SHUTTER", "BG_ICE_SHUTTER"),
+	("ACTOR_EN_KAKASI2", "EN_KAKASI2", "EN_KAKASI2"),
+	("ACTOR_EN_KAKASI3", "EN_KAKASI3", "EN_KAKASI3"),
+	("ACTOR_OCEFF_WIPE4", "OCEFF_WIPE4", "OCEFF_WIPE4"),
+	("ACTOR_EN_EG", "EN_EG", "EN_EG"),
+	("ACTOR_BG_MENKURI_NISEKABE", "BG_MENKURI_NISEKABE", "BG_MENKURI_NISEKABE"),
+	("ACTOR_EN_ZO", "EN_ZO", "EN_ZO"),
+	("ACTOR_OBJ_MAKEKINSUTA", "OBJ_MAKEKINSUTA", "OBJ_MAKEKINSUTA"),
+	("ACTOR_EN_GE3", "EN_GE3", "EN_GE3"),
+	("ACTOR_OBJ_TIMEBLOCK", "OBJ_TIMEBLOCK", "OBJ_TIMEBLOCK"),
+	("ACTOR_OBJ_HAMISHI", "OBJ_HAMISHI", "OBJ_HAMISHI"),
+	("ACTOR_EN_ZL4", "EN_ZL4", "EN_ZL4"),
+	("ACTOR_EN_MM2", "EN_MM2", "EN_MM2"),
+	("ACTOR_BG_JYA_BLOCK", "BG_JYA_BLOCK", "BG_JYA_BLOCK"),
+	("ACTOR_OBJ_WARP2BLOCK", "OBJ_WARP2BLOCK", "OBJ_WARP2BLOCK"),
 ]
 
 ootEnumLinkIdle = [
@@ -512,6 +512,7 @@ ootEnumEmptyType = [
 	('Water Box', 'Water Box', 'Water Box'),
 	('Cull Group', 'Cull Group', 'Cull Group'),
 	('LOD', 'LOD Group', 'LOD Group'),
+	('Cutscene', 'Cutscene', 'Cutscene'),
 	#('Camera Volume', 'Camera Volume', 'Camera Volume'),
 ]
 
@@ -1412,6 +1413,12 @@ ootEnumSceneSetupPreset = [
 	("All Non-Cutscene Scene Setups", "All Non-Cutscene Scene Setups", "All Non-Cutscene Scene Setups"),
 ]
 
+ootEnumCSWriteType = [
+	("Custom", "Custom", "Provide the name of a cutscene header variable"),
+	("Embedded", "Embedded", "Cutscene data is within scene header (deprecated)"),
+	("Object", "Object", "Reference to Blender object representing cutscene")
+]
+
 ootEnumCSListType = [
 	("Textbox", "Textbox", "Textbox"),
 	("FX", "Scene Trans FX", "Scene Trans FX"),
@@ -1426,8 +1433,8 @@ ootEnumCSListType = [
 ]
 
 ootEnumCSListTypeIcons = [
-    'ALIGN_BOTTOM', 'COLORSET_10_VEC', 'LIGHT_SUN', 'TIME', 'PLAY', 'SNAP_FACE', 'IPO_EASE_IN_OUT',
-    'OPTIONS', 'EVENT_F9', 'QUESTION'
+	'ALIGN_BOTTOM', 'COLORSET_10_VEC', 'LIGHT_SUN', 'TIME', 'PLAY', 'SNAP_FACE', 'IPO_EASE_IN_OUT',
+	'OPTIONS', 'EVENT_F9', 'QUESTION'
 ]
 
 ootEnumCSListTypeListC = {
@@ -1458,18 +1465,18 @@ ootEnumCSListTypeEntryC = {
 
 ootEnumCSTextboxType = [
 	("Text", "Text", "Text"),
-    ("None", "None", "None"),
-    ("LearnSong", "Learn Song", "Learn Song")
+	("None", "None", "None"),
+	("LearnSong", "Learn Song", "Learn Song")
 ]
 
 ootEnumCSTextboxTypeIcons = [
-    'FILE_TEXT', 'HIDE_ON', 'FILE_SOUND'
+	'FILE_TEXT', 'HIDE_ON', 'FILE_SOUND'
 ]
 
 ootEnumCSTextboxTypeEntryC = {
-    "Text": 'CS_TEXT_DISPLAY_TEXTBOX',
-    "None": 'CS_TEXT_NONE',
-    "LearnSong": 'CS_TEXT_LEARN_SONG',
+	"Text": 'CS_TEXT_DISPLAY_TEXTBOX',
+	"None": 'CS_TEXT_NONE',
+	"LearnSong": 'CS_TEXT_LEARN_SONG',
 }
 
 ootEnumCSTransitionType = [
